rename default config file to .sample
The general plan is so we can still happily update with git.
The defaults in this config file should be hard-coded into the
daemons and hooks, so that it still run when the file is named .sample
while providing the possibility to override these defaults.

Speaking of which: How do I override defaults from gitpubsub.lua?!